Add the `color` and `file` property types
Added a `bool` property type in the union, but it's still an `int`. Code
made with TMX 0.16.0 do not need to update because it's an union, you
can still use ->integer to get the boolean value.

Added a `color` property type stored in an unsigned int (AARRGGBB).

Added a `file` property type stored in a char*, the file is not loaded
in memory.

Added a PT_COLOR and a PT_FILE entry as possible values for
tmx_property->type.